opentrend/requirements.txt
# OpenTrend - Fashion Trend Forecasting System
# Requirements

# Core
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Computer Vision
torch>=1.10.0
torchvision>=0.11.0
ultralytics>=8.0.0
Pillow>=8.0.0

# Data Ingestion
pytrends>=4.9.0

# Machine Learning
scikit-learn>=0.24.0

# Forecasting
prophet>=1.1.0

# API (optional)
fastapi>=0.68.0
uvicorn>=0.15.0

# Visualization
matplotlib>=3.4.0
